fix: resolve all frontend ESLint errors
