+++ booking_system_frontend/src/hooks/useUserContext.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react';
+import type { UserContextType } from '../types';
+
+export const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export const useUser = (): UserContextType => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error('useUser must be used within a UserProvider');
+  }
+  return context;
+};
