+++ booking_system_frontend/src/services/api.ts
@@ -206,9 +206,13 @@ export const releaseHold = async (holdId: string): Promise<Hold> => {
  * Check if response is an error
  */
 export const isErrorResponse = (
-  response: any
+  response: unknown
 ): response is ErrorResponse => {
-  return response && response.success === false;
+  return (
+    typeof response === 'object' &&
+    response !== null &&
+    (response as ErrorResponse).success === false
+  );
 };
 
 /**
